Organize all voice util files under new utils/voice folder
